LMN-1060 Refactored interface to split image categories and chip categories
